DRBN_SKF/src/requirements.txt
backcall==0.2.0
charset-normalizer==2.1.1
cycler==0.11.0
decorator==5.1.1
docopt==0.6.2
einops==0.5.0
fonttools==4.37.4
glob2==0.7
idna==3.4
imageio==2.22.1
imutils==0.5.4
ipython==7.34.0
jedi==0.18.1
kiwisolver==1.4.4
lpips==0.1.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
natsort==8.2.0
networkx==2.6.3
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pyclipper==1.3.0.post3
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.2.1
six==1.16.0
tifffile==2021.11.2
torch==0.10.1+cu111
torchaudio==0.10.1+rocm4.1
torchvision==0.11.2+cu111
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
utils==1.0.1
wcwidth==0.2.5
yacs==0.1.8
yarg==0.1.9
